pyproject.toml
[project]
name = "cot-prompting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.117.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "pypdf>=6.1.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "rq>=2.6.0",
    "sentence-transformers>=5.1.1",
    "uvicorn>=0.37.0",
]
